+++ Gui/GUI.py
@@ -3,12 +3,9 @@ from tkinter import filedialog, ttk
 from tkcalendar import DateEntry
 from datetime import date
 from tkinter import messagebox
-from Gui.GuiHelperService import center_window
-from Gui.consts import CONST_NOT_ALL_FIELDS_FILLED
-from Gui.consts import sensor_type_list
-
-
-
+from Gui.center_window import center_window
+from consts import CONST_NOT_ALL_FIELDS_FILLED
+from consts import sensor_type_list
 
 
 def pick_save_folder():
